+++ src/services/NodeMonitor.ts
@@ -12,9 +12,9 @@ import { AbstractTimeSeries, AbstractTimeSeriesDocument } from '@src/models/Abst
 import { memoryCache } from '@src/services/MemoryCache';
 import { Logger } from '@src/infrastructure';
 
-import { INode } from '@src/models/Node';
+import { INode, validateNodeModel } from '@src/models/Node';
 import { symbol, monitor } from '@src/config';
-import { isAPIRole, isPeerRole, getNodeURL, basename } from '@src/utils';
+import { isAPIRole, isPeerRole, getNodeURL, basename, splitArray, sleep } from '@src/utils';
 
 const logger: winston.Logger = Logger.getLogger(basename(__filename));
 
@@ -24,6 +24,9 @@ export class NodeMonitor {
 	private nodeList: INode[];
 	private isRunning: boolean;
 	private interval: number;
+	private nodeInfoChunks: number;
+	private nodeInfoDelay: number;
+	private networkIdentifier: number;
 
 	constructor(_interval: number) {
 		this.nodesStats = new NodesStats();
@@ -35,6 +38,9 @@ export class NodeMonitor {
 		this.nodeList = [];
 		this.isRunning = false;
 		this.interval = _interval || 300000;
+		this.nodeInfoChunks = 500;
+		this.nodeInfoDelay = 1000;
+		this.networkIdentifier = 152; // default Testnet
 		this.cacheCollection();
 	}
 
@@ -48,6 +54,8 @@ export class NodeMonitor {
 		try {
 			this.isRunning = true;
 			this.clear();
+
+			await this.getNetworkType(); // Read network Type from provided nodes.
 
 			await this.getNodeList();
 			await this.getNodeListInfo(); //HostInfo.getInfoForListOfNodes(this.nodeList);
@@ -73,11 +81,9 @@ export class NodeMonitor {
 	private getNodeList = async (): Promise<any> => {
 		// Init fetch node list from config nodes
 		logger.info(`Getting node list`);
-		let counter = 0;
 
 		for (const nodeUrl of symbol.NODES) {
-			counter++;
-			const peers = await this.fetchNodesByURL(nodeUrl);
+			const peers = await this.fetchNodesByURL(nodeUrl, true);
 
 			this.addNodesToList(peers);
 		}
@@ -97,22 +103,50 @@ export class NodeMonitor {
 		return Promise.resolve();
 	};
 
-	private fetchNodesByURL = async (nodeUrl: string): Promise<Array<INode>> => {
+	private fetchNodesByURL = async (nodeUrl: string, includeCurrent: boolean = false): Promise<Array<INode>> => {
+		let nodeList = [];
+
+		if (includeCurrent) {
+			try {
+				const nodeInfo = await HTTP.get(nodeUrl + '/node/info', {
+					timeout: monitor.REQUEST_TIMEOUT,
+				});
+				const host = new URL(nodeUrl).hostname;
+
+				nodeList.push({
+					...nodeInfo.data,
+					host,
+				});
+			} catch (e) {
+				logger.error(`FetchNodesByURL. Failed to get /node/info from "${nodeUrl}". ${e.message}`);
+			}
+		}
+
 		try {
-			const nodeList = await HTTP.get(nodeUrl + '/node/peers', {
+			const nodePeers = await HTTP.get(nodeUrl + '/node/peers', {
 				timeout: monitor.REQUEST_TIMEOUT,
 			});
 
-			if (Array.isArray(nodeList.data)) return nodeList.data;
-		} catch (e) {}
-		return [];
+			if (Array.isArray(nodePeers.data)) nodeList = [...nodeList, ...nodePeers.data];
+		} catch (e) {
+			logger.error(`FetchNodesByURL. Failed to get /node/peers from "${nodeUrl}". ${e.message}`);
+		}
+
+		return nodeList;
 	};
 
 	private getNodeListInfo = async () => {
-		logger.info(`Getting node info for ${this.nodeList.length} nodes`);
-		const nodeInfoPromises = this.nodeList.map(this.getNodeInfo);
+		logger.info(`Getting node info total for ${this.nodeList.length} nodes`);
+		const nodeInfoPromises = [...this.nodeList].map(this.getNodeInfo);
+		const nodeInfoPromisesChunks = splitArray(nodeInfoPromises, this.nodeInfoChunks);
 
-		this.nodeList = await Promise.all(nodeInfoPromises);
+		this.nodeList = [];
+
+		for (const chunk of nodeInfoPromisesChunks) {
+			logger.info(`Getting node info for chunk of ${chunk.length} nodes`);
+			this.addNodesToList((await Promise.all(chunk)) as INode[]);
+			await sleep(this.nodeInfoDelay);
+		}
 		this.nodeList.forEach((node) => this.nodesStats.addToStats(node));
 	};
 
@@ -134,7 +168,7 @@ export class NodeMonitor {
 
 			if (nodeWithInfo.publicKey) nodeWithInfo.rewardPrograms = await NodeRewards.getNodeRewardPrograms(nodeWithInfo.publicKey);
 		} catch (e) {
-			logger.error(`failed to get info. ${e.message}`);
+			logger.error(`GetNodeInfo. Failed to fetch info for "${node}". ${e.message}`);
 		}
 
 		return nodeWithInfo;
@@ -179,9 +213,31 @@ export class NodeMonitor {
 		}
 	};
 
+	private getNetworkType = async (): Promise<void> => {
+		for (const nodeUrl of symbol.NODES) {
+			const url = new URL(nodeUrl);
+
+			const nodeInfo = await ApiNodeService.getNodeInfo(url.hostname, Number(url.port || monitor.API_NODE_PORT));
+
+			if (nodeInfo) {
+				this.networkIdentifier = nodeInfo.networkIdentifier;
+				logger.info(`Found network identifier ${nodeInfo.networkIdentifier}`);
+				return;
+			}
+		}
+
+		logger.info(`Network identifier not found in ${symbol.NODES}, using default ${this.networkIdentifier}`);
+	};
+
 	private addNodesToList = (nodes: INode[]) => {
 		nodes.forEach((node: INode) => {
-			if (!!this.nodeList.find((addedNode) => addedNode.publicKey === node.publicKey)) return;
+			if (
+				node.networkIdentifier !== this.networkIdentifier ||
+				!!this.nodeList.find((addedNode) => addedNode.publicKey === node.publicKey) ||
+				!validateNodeModel(node)
+			)
+				return;
+
 			this.nodeList.push(node);
 		});
 	};