+++ CHANGELOG.md
@@ -3,6 +3,22 @@ All notable changes to this project will be documented in this file.
 
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [v1.0.1][v1.0.1] - 25-May-2021
+
+### Milestone: [catapult-server@v1.0.0.0][catapult-server@v1.0.0.0]
+
+Package  | Version  | Link
+---|---|---
+REST Core| v2.3.6 | [catapult-rest][catapult-rest@v2.3.6]
+SDK Core| v1.0.1 | [symbol-sdk][symbol-sdk@v1.0.1]
+
+- Added the network type check to the Node Monitor Service [#53](https://github.com/nemgrouplimited/symbol-statistics-service/issues/53)
+- Added version of REST to the Node List [#55](https://github.com/nemgrouplimited/symbol-statistics-service/issues/53)
+- Fixed the peer node status check issue [#58](https://github.com/nemgrouplimited/symbol-statistics-service/pull/58)
+- Improved DB collection store [#58](https://github.com/nemgrouplimited/symbol-statistics-service/pull/58)
+- Included the initial nodes from config to the list [#58](https://github.com/nemgrouplimited/symbol-statistics-service/pull/58)
+- Added missing utf8 dependency
+
 ## [v1.0.0][v1.0.0] - 14-Mar-2021
 
 ### Milestone: [catapult-server@v1.0.0.0][catapult-server@v1.0.0.0]
@@ -71,7 +87,8 @@ REST Core| v2.1.0 | [catapult-rest](https://github.com/nemtech/catapult-rest/rel
 ### Fixes
 - Cors error. [#13](https://github.com/nemgrouplimited/symbol-statistics-service/issues/13)
 
-
+[v1.0.1]: https://github.com/nemfoundation/symbol-statistics-service/releases/tag/v1.0.1
+[v1.0.0]: https://github.com/nemfoundation/symbol-statistics-service/releases/tag/v1.0.0
 [v0.3.0]: https://github.com/nemfoundation/symbol-statistics-service/releases/tag/v0.3.0
 [v0.2.0]: https://github.com/nemfoundation/symbol-statistics-service/releases/tag/v0.2.0
 [v0.1.0]: https://github.com/nemfoundation/symbol-statistics-service/releases/tag/v0.1.0
@@ -83,6 +100,8 @@ REST Core| v2.1.0 | [catapult-rest](https://github.com/nemtech/catapult-rest/rel
 [symbol-sdk@v0.23.3-alpha]: https://www.npmjs.com/package/symbol-sdk/v/0.23.3-alpha-202102181227
 [symbol-sdk@v0.23.3]: https://www.npmjs.com/package/symbol-sdk/v/0.23.3
 [symbol-sdk@v1.0.0]: https://www.npmjs.com/package/symbol-sdk/v/1.0.0
+[symbol-sdk@v1.0.1]: https://www.npmjs.com/package/symbol-sdk/v/1.0.1
 
 [catapult-rest@v2.3.3]: https://github.com/nemtech/catapult-rest/releases/tag/v2.3.3
 [catapult-rest@v2.3.4]: https://github.com/nemtech/catapult-rest/releases/tag/v2.3.4
+[catapult-rest@v2.3.6]: https://github.com/nemtech/catapult-rest/releases/tag/v2.3.6